change(web): moves uiTouchHandlerBase class
